Add drink size selector functionality
Since we have a bit of a strange system where the drink size is
indicated as an "ingredient", all we need to do is convert the size that
the user selected into a customization in the frontend itself. No
backend API changes are required since the customization automatically
gets calculated as part of the price.